Address reviewer feedback in PR comment: Changed the file name to "parameters.py" to "params.py" and did refactor.
