+++ src/pythermalcomfort/models.py
@@ -30,7 +30,7 @@ from pythermalcomfort.jos3_functions import matrix
 from pythermalcomfort.jos3_functions.matrix import NUM_NODES, INDEX, VINDEX, BODY_NAMES, remove_body_name
 from pythermalcomfort.jos3_functions import construction as cons
 from pythermalcomfort.jos3_functions.construction import (_BSAst, _to17array)
-from pythermalcomfort.jos3_functions.params import ALL_OUT_PARAMS
+from pythermalcomfort.jos3_functions.parameters import ALL_OUT_PARAMS
 from pythermalcomfort.__init__ import __version__
